fix(format): 32-bit narrowing in read_library (uint64 tocOffset -> size_t)
On armv7/wasm32 (size_t is 32-bit) the Reader brace-init narrowed the 64-bit
tocOffset to size_t under -Werror -Wc++11-narrowing. It is validated <= buffer
size, so cast explicitly. Desktop (64-bit size_t) didn't surface it.

+++ source/vshadersystem/src/vsh_format.cpp
@@ -595,7 +595,8 @@ namespace vshadersystem::v1
             return R::err({ErrorCode::eDeserializeError, "bad TOC offset"});
 
         Library lib;
-        Reader  tr{r.p, r.n, tocOffset, false};
+        // tocOffset is validated <= r.n above, so it fits size_t (32-bit safe).
+        Reader  tr{r.p, r.n, static_cast<size_t>(tocOffset), false};
         for (uint32_t i = 0; i < entryCount; ++i)
         {
             LibraryEntry e;
